Changing up how processors work to make it a bit easier to add to them.
