+++ packages/string-templates/src/processors/postprocessor.js
@@ -1,6 +1,10 @@
+const { FIND_HBS_REGEX } = require("../utilities")
+
 class Postprocessor {
   constructor(name, fn) {
     this.name = name
     this.fn = fn
   }
 }
+
+module.exports.processors = []
